PID and code clean up
Removed some files that were not being used and added new files to be used with PID. Also updated and tuned the PID but the arm PID still needs to be tested to be safe. Final note the next update will be a reword of the swerve drive system since we are switching to REV Robotics's swerve template.